refactor: move target name to mouse manager
