--- Assets/Scripts/TargetName.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4bd07797abbc8754eb87192f1fb7a004
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
